+++ res/values/strings.xml
@@ -12007,7 +12007,7 @@
     <!-- If blurs are supported on SurfaceFlinger. [CHAR LIMIT=60] -->
     <string name="enable_blurs_on_windows_title">Enable blurs</string>
     <!-- If blurs are supported on SurfaceFlinger, summary. [CHAR LIMIT=NONE] -->
-    <string name="enable_blurs_on_windows_summary">Enables window blurs at compositor level. Requires device reboot.</string>
+    <string name="enable_blurs_on_windows_summary">Enables blurs in some UI elements; requires device reboot</string>
 
     <!-- Title for the top level Privacy Settings [CHAR LIMIT=30]-->
     <string name="privacy_dashboard_title">Privacy</string>
