+++ res/xml/development_settings.xml
@@ -529,11 +529,6 @@
             android:summary="@string/hwui_force_dark_summary" />
 
         <SwitchPreference
-            android:key="enable_blurs_on_windows"
-            android:title="@string/enable_blurs_on_windows_title"
-            android:summary="@string/enable_blurs_on_windows_summary" />
-
-        <SwitchPreference
             android:key="force_msaa"
             android:title="@string/force_msaa"
             android:summary="@string/force_msaa_summary" />
