+++ res/xml/display_settings.xml
@@ -145,6 +145,11 @@
         android:title="@string/tap_to_wake"
         android:summary="@string/tap_to_wake_summary" />
 
+    <SwitchPreference
+        android:key="enable_blurs_on_windows"
+        android:title="@string/enable_blurs_on_windows_title"
+        android:summary="@string/enable_blurs_on_windows_summary" />
+
     <ListPreference
         android:key="theme"
         android:title="@string/device_theme"
